Se separa el modulo del lexer del principal. Mas facil de entender.
